feat: prestador confirmar a realizacao do servico

+++ app/controllers/user_services_controller.rb
@@ -56,24 +56,36 @@ class UserServicesController < ApplicationController
   def destroy
     service_finish = user_service_delete
     service = Service.find(service_finish[:service_id])
-    service.finished = true
-    service.save
+    prestador_confirm = ActiveModel::Type::Boolean.new.cast(service_finish[:prestador_confirm])
+    if prestador_confirm
+      service.finished_prestador = true
+      service.save
 
-    messages = Message.where(chat_id: service_finish[:chat])
-    messages.destroy_all
+      flash[:notice] = 'Serviço Marcado como Realizado!'
+      redirect_to user_path(service.user_selected_id)
+    elsif !prestador_confirm && service.finished_prestador
+      service.finished = true
+      service.save
 
-    user = User.find(service.user_selected_id)
-    user.services_realized += 1
+      messages = Message.where(chat_id: service_finish[:chat])
+      messages.destroy_all
 
-    if user.score.nil?
-      user.score = service_finish[:comment].to_i
+      user = User.find(service.user_selected_id)
+      user.services_realized += 1
+
+      if user.score.nil?
+        user.score = service_finish[:comment].to_i
+      else
+        user.score += service_finish[:comment].to_i
+      end
+      user.save
+
+      flash[:notice] = 'Serviço Finalizado'
+      redirect_to user_path(service.user)
     else
-      user.score += service_finish[:comment].to_i
+      flash[:alert] = 'O prestador precisa confirmar a realização do serviço!'
+      redirect_to user_service_path
     end
-    user.save
-
-    flash[:notice] = 'Serviço Finalizado'
-    redirect_to user_path(service.user)
   end
 
   private
@@ -91,6 +103,6 @@ class UserServicesController < ApplicationController
   end
 
   def user_service_delete
-    params.require(:user_service).permit(:comment, :chat, :service_id)
+    params.require(:user_service).permit(:comment, :chat, :service_id, :prestador_confirm)
   end
 end
